pyproject.toml
[project]
name = "migrate"
version = "0.0.2"
description = "GitHub migration helper"
requires-python = ">=3.11"
dependencies = [
    "click~=8.1.3",
    "fastcore~=1.5.27",
    "ghapi~=1.0.3",
    "pyaml~=21.10.1",
    "pytest~=7.2.0",
    "pynacl~=1.5.0",
    "requests~=2.28.1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
testpaths = "tests"
pythonpath = [
    "migrate"
]

[tool.black]
line-length = 90

[project.optional-dependencies]
dev = [
    "black>=22.3.0",
    "flake8>=4.0.1",
    "setuptools>=65.6.3"
]

[project.scripts]